Creation of InstructionRepository

+++ server/src/main/java/server/database/InstructionRepository.java
@@ -4,6 +4,6 @@ import commons.Instruction;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
-public interface InstructionRepository extends JpaRepository<Instruction, Integer> {
+public interface InstructionRepository extends JpaRepository<Instruction, Long> {
 
 }
